refactor(chunk-store): rename ports.ts to types.ts (#1189)
File contains only type/interface definitions (ResponseLike, HashPort,
CachePort, etc.) but was named ports.ts instead of types.ts.

Rollback: git revert <hash> --no-edit
