+++ packages/chunk-store/src/adapters/NobleSha3HashPort.ts
@@ -1,5 +1,5 @@
 import { sha3_256 } from '@noble/hashes/sha3';
-import type { HashAlgorithm, HashPort } from '~/ports';
+import type { HashAlgorithm, HashPort } from '~/types';
 
 export class NobleSha3HashPort implements HashPort {
   digest(buffer: ArrayBuffer, algo: HashAlgorithm): string {
